+++ test/src/unit-tile-metadata.cc
@@ -737,7 +737,7 @@ struct CPPVarTileMetadataFx {
     values.reserve(num_cells_);
     uint64_t var_size = 0;
     for (uint64_t i = 0; i < num_cells_; i++) {
-      values.emplace_back(rand() % num_strings);
+      values.emplace_back(static_cast<int>(rand() % num_strings));
       var_size += strings_[values.back()].size();
     }
